Much trait re-naming
