+++ p2panda-auth/src/test_utils/network.rs
@@ -8,7 +8,7 @@ use rand::seq::SliceRandom;
 
 use crate::Access;
 use crate::group::{GroupAction, GroupControlMessage, GroupMember};
-use crate::traits::{GroupStore, Operation, Ordering};
+use crate::traits::{GroupStore, Operation, Orderer};
 
 use super::{
     GroupId, MemberId, MessageId, TestGroup, TestGroupState, TestGroupStore, TestOperation,
