+++ p2panda-auth/src/traits/mod.rs
@@ -4,18 +4,16 @@ use std::fmt::Debug;
 use std::hash::Hash as StdHash;
 
 mod dgm;
-mod store;
 mod operation;
-mod ordering;
-mod query;
+mod orderer;
 mod resolver;
+mod store;
 
-pub use dgm::GroupMembership;
-pub use store::GroupStore;
+pub use dgm::{Group, GroupMembership};
 pub use operation::Operation;
-pub use ordering::Ordering;
-pub use query::GroupMembershipQuery;
+pub use orderer::Orderer;
 pub use resolver::Resolver;
+pub use store::GroupStore;
 
 /// Handle to identify a group member.
 ///
