+++ p2panda-auth/src/traits/orderer.rs
@@ -6,7 +6,7 @@ use std::error::Error;
 /// Interface for processing messages which have particular ordering requirements.
 ///
 /// Messages have an author id, a unique identifier and a generic payload.
-pub trait Ordering<ID, OP, P> {
+pub trait Orderer<ID, OP, P> {
     type State;
     type Operation: Operation<ID, OP, P>;
     type Error: Error;
